feat: remove ipfs.io dns RPC calls
This also refactors `dns-resolvers/default.ts` so that the recursive
logic of the `defaultResolver` will be shared between browser and nodejs
consumers.
