+++ packages/ipns/package.json
@@ -166,6 +166,7 @@
     "@libp2p/logger": "^3.0.2",
     "@libp2p/peer-id": "^3.0.2",
     "@libp2p/record": "^3.0.0",
+    "dns-over-http-resolver": "^2.1.3",
     "dns-packet": "^5.6.0",
     "hashlru": "^2.3.0",
     "interface-datastore": "^8.0.0",
@@ -185,7 +186,7 @@
     "sinon-ts": "^1.0.0"
   },
   "browser": {
-    "./dist/src/dns-resolvers/default.js": "./dist/src/dns-resolvers/default.browser.js"
+    "./dist/src/dns-resolvers/resolver.js": "./dist/src/dns-resolvers/resolver.browser.js"
   },
   "typedoc": {
     "entryPoint": "./src/index.ts"
